scale icon with DPI
Windows seems to have worse downscaling than Inkscape, so different
sizes of the logo will be used at different DPIs instead of passing
a large icon at all times.

Only Windows has a separate API for setting the taskbar icon, but it
will be twice the resolution up to 128x128 on Windows.