implement sidebar, misc changes
Made an initial implementation of the sidebar, which contains a list of
open REPLs and REPLs that can be created. Added more fonts to the text
renderer API for rendering the sidebar (and other non-editor UI elements
in the future).

Set a reasonable-seeming default window size of 1400x900, and a minimum
size of 500x300. Moved BoundingBox to its own module as its API is
repetitive and cluttered up the drawable module.