draw panes, prepare for drawing icons, include example interface...
Added a correct Drawable implementation for Panes, so panes are now
drawn. Pane still has an empty implementation, so no content is drawn,
only the layout of the interface.

Added a UserData type to Drawable used for types that need extra
information to be drawn, like whether a pane is focused or not.

Moved the logo files to their own folder to keep the icons folder
organized as more icons are added. Added an icons/ui folder for icons
used in the user interface, like the close icon for closing REPLs.

Moved icon logic to its own module, including an Icons struct that will
handle creation of both winit icons and wgpu textures for required
icons, keeping track of dark/light mode, scale factor, etc.

Added a dummy interface in App::new to demonstrate the interface layout
and design.

Added a few colors to the configuration for the newly implemented UI
features.

Added a Theme enum with Dark and Light variants to support dark/light
themes.