render icons, restructure
Added a rendering pipeline for textures, along with the associated
vertex and index buffers, texture, shaders, and render pass.

Added atlas creation and texture filling to Icons.

Added a draw_icon_rect method to DrawingContext for drawing icons within
the interface.

Added code to draw close icons on tabs.

Simplified the format of icon modules.

Moved the rendering pipelines, bind groups, and texture to the Pipelines
struct.

Replaced the channel-per-list setup previously used to send the vertex
and index lists back to the renderer with a single channel and a
ListMessage discriminator.

Fixed the build.rs file to correctly recompile when shaders are changed.