update dependencies, cleanup
Update wgpu to 0.5 and wgpu_glyph to 0.9. Use a published alpha version
of tinyvec instead of using a git repository.

wgpu 0.5 doesn't depend on zerocopy, so it has been replaced by bytemuck
which is more straightforward. Added pollster to block on wgpu futures.
Added an explicit dependency on glyph_brush for types that wgpu_glyph
doesn't re-export.

winit 0.22.2 fixed ControlFlow::Poll, so the temporary workaround was
removed.

Simplified Action, removing the duplication of a None variant in both
the outer Action layer and the inner ActionData layer.

With wgpu 0.5's async interface based on futures, some data shuttling
code was removed and replaced with an interface on DrawingManager for
iterating over every vertex/index. This is made possible because the
borrow checker can analyze through async blocks better than it could
through the callbacks that the previous interface used.